Virtual Keyboard other than QWERTY ?
On Tue, Dec 25, 2012 at 4:25 PM, oc-spam65 wrote:
> Hello,
>
> Is it possible to use Yoshimi with an AZERTY computer keyboard?
> How can it be managed? It seems that only QWERTY and Dvorak are available.
>
> Thank you,
>
> O.C.
>
> Details: Debian Testing / JACK / Yoshimi 0.060.12 / AZERTY keyboard

Zynaddsubfx supports azerty since v. 2.4.1 Pre (May 2, 2010)
http://zynaddsubfx.sourceforge.net/news.html
Yoshimi is based on 2.4.0, so it missed that one.

The relevant commit to zyn:
http://zynaddsubfx.git.sourceforge.net/git/gitweb.cgi?p=zynaddsubfx/zynaddsubfx;a=commit;h=b6074bdb7b50ca0b75e7e4a426e1eabbb324176b

Blindly applying the same changes to the yoshimi-source happens to work,
patch attached.

to apply it:
$ wget http://downloads.sourceforge.net/project/yoshimi/1.0/yoshimi-1.0.0.tar.bz2
$ tar xvjf yoshimi-1.0.0.tar.bz2
$ cd yoshimi-1.0.0/src/
$ patch -p2 -i some/path/to/yoshimi_azerty.patch

then build, e.g.:
$ cmake .
$ make
and run it
$ ./yoshimi
or install

merry christmas.
 -d