Multiple audio outputs
On 1/27/13, Johannes Mueller <johmue@users.sourceforge.net> wrote:
> Can I get to have two audio (one stereo) outputs for every "enabled" part?

Yes. I eventually remembered that things can be easier to handle from
the command line, so I skipped over all the UI and XML worries and
major modifications. This patch against the current 1.0 branch makes a
few assumptions and adds three new options:

-M or --jack-multi will switch on per-part output ports.

-B <value> or --jack-multi-mask=<value> takes a bitmask to let you
optimize for your needs by enabling only some extra outputs at
startup. There's no way to change the number of ports in a single
run... that makes things easy and lets me finish in a few hours. Give
it a number like 0x0003 and only parts 1 and 2 will have ports, 0xE000
for only on parts 14, 15, and 16. Of course it's 0xFFFF (65535) by
default. It uses an unsigned long int so I suppose it's good for 64 if
you edited your CMakeLists.txt to increase NUM_MIDI_PARTS. I don't
know about your CPU. I listened to the ports on part 46 out of 48 and
it worked but that was by itself.

-P <text> or --master-prefix=<text> will prepend text to the master
out channels so you can break or fix the port listings at will. The
default is of course nothing extra, just "left" and "right".

The individual output ports are also "left" and "right", prefixed with
_NN_ so they'll start below letters in an alphabetical list, and if
you leave the main outputs' names alone, they stay on top. NN is the
zero-padded part number starting at 1, just like parts.

The part output buffers are taken after insertion effects. I did that
without knowing which buffer to read from but I left it that way
because it seems to follow the rule of least surprise, if that's a
thing...

It's not really tested, much. I got it to stop segfaulting and made
sure the tabs were spaces and shipped it. I disclaim everything except
that it runs over here, for now. good luck, have fun :)

--
Rob

Signed-off-by: Jeremy <jeremy@piertje.dekenwaarestraat.nl>