Yoshimi NSM patch port, partially working
Thanks goes to Jonathan M. Liles for the original patch,
without which this would not have been acomplished.

There are a few // NSM_TODO comments left around the source,
those areas *definatly* need attention, and generally the
whole patch could do with a review.

The code in main.cpp is tempermental - the order in which
the SynthEngine and NSM clients are created is very important
to NSM, and generally not so for Yoshimi (i think..) so its
probably not an ideal solution. Anyway :)

Toodles, -Harry van Haaren