Refactoring: disentangle the implementation
Up to now, the implementation had various intermingled conditional compilation cases,
which made reasoning about the implementation difficult.

This commit changes no functionality, but reorganises the code into a dedicated
implementation class for each of the possible PRNG implementations. The conditional
compilation is now concentrated at a single location, where we define a type alias
to the actual implementation to be used.

- in this form, it is now simple to add further implementation algorithms
- moreover, this could be the foundation to make this choice switchable from
  the instrument definition, with zero overhead (using templates)