Include a literal copy from the Glibc 2.28 random_r() implementation
This adds yet another PRNG implementation, which incorporates a literal copy
from the Glibc 2.28 implementation, without any further simplifications.

I use Glibc 2.28 here, since it adds a simplification/clarification for the
rightshift to drop the least random bit; Since newer C++ standards require
to use an arithmetic right shift on signed values, the internal variable val
is now defined to be uint32_t, which also allows to omit the factor  & 0x7fffffff;
used in previous versinos of the Glibc.

TODO: This changeset includes a huge pile of constants and code
which can be simplified, given that our state array is known to be inline and 256 bytes large