switch the TrinomaialPRNG algorithm to use uint32_t throughout
In fact this does not make any difference on the level of the binary numbers,
and the compiler emits the same assembly.

Yet it makes the code way more straight forward, and avoids pitfalls, like e.g.
with right shift (which is arithmetic right shift for signed numbers, while we
want the classic mere bitshift).

However, we retain the seed calculation as-is, to avoid changing behaviour
on INT_MAX < seed < UINT_MAX