clean-up: PRNG init can not fail
The library version has a failure exit in case the user proivdes a NULL ptr.
Since we always pass a pointer to our PRNG state embedded in the class,
the init() function can not fail. Only the allocation of the class storage,
which would raise an std::exception anyway.

Also added some comments and reinstated the rot() function from the original
Jenkins implementation, since this makes the code more legible (and the compiler
will inline anyway, verified that by looking into the assembly)

** Happy New Year **