Cross-check and compare my PRNG refactoring with Will's improvements
While working on the PRNG, Will did some further improvements and bugfixes.
This changeset reinstates these changes, adapted to the recent refactorings.

+ Initially seed the global PRNG from current time, which makes this initial seed truely random.
  Previously the code used 'samplerate + buffersize + oscilsize', which made
  the initial seed dependent on some arbitrary contextual information, but didn't
  make it truely random.

+ Moved this single global reseeding from SynthEngine::Init() up into the SynthEngine-ctor.
  This changes no functionality, but makes the code more obvious. Init() is always
  called immediately after creating the global SynthEngine instance.

- However, I did not reinstate any of the changes to the reseed mechanism.
  Close investigation of the call graphs indicate these changes to be far more
  intricate as it might look on first sight: The OscilGen::get() function
  happens from within the NoteOn-Callgraph, and thus re-randomises the
  PRNGs shared in the OscilGen instances per Kititem-component.
  Moreover, changing the reseeding this way would also interfer with the
  global Hrandgrouping parameter, especially when using another oscil within
  the same AddSynth for modulation.

-----------
For context: while RC testing, subtle sonic changes to some complex layered patches
were identified to be related to the PRNG reworkings.

For further investigation and refactoring I then bulk reverted the five
relevant changes from Will, which introduced the Jenkins PRNG algorithm.

Subsequently we resorted to using the random_r() algorithm again, but we
extracted it from Glibc 2.28 and maintain it now in-tree, in order to
guarantee compatibility with previously created instruments/patches