complete the switch to Trinomial PRNG, extracted from Glibc-2.28
After spotting some sonic changes while RC testing, we weren't able
to identify any obvious flaw in the Jenkins PRNG. My extensive
measurements and listening comparisons rather indicate a different
fluctuation patterning of the Jenkins PRNG algo as reason, when
compared with the PRNG of the C standard lib, which Yoshimi
used to rely on.

However, dependency on an external library for such a crucial component
is a concern; just linking to some other platform C library might cause
similar subtle sonic changes. For this reason, we decided to extract
the actual implementation from GNU Glibc-2.28 and incorporate it in-tree.

This refactoring is now complete and verified.
Drop the dependency on external random_r() implementation.