Do not manipulate the buffersize (Alsa)
The Alsa-Engine used to manipulate the Buffersize in the Configuration,
when, after opening the interface, the ALSA framework reported back a different
*output* buffersize than initially requested.

This is altogether unnecessary, since all output frameworks are able to
deal with an output buffer size different than the *calculation* buffersize
of SynthEngine.

Config::Buffersize (which is written to XML), SynthEngine::buffersize
and the output buffersize are three distinct concepts and should be kept apart.