Output: adapt LV2 plugin to also use the same processing pattern as Jack/ALSA
This changeset rearranges calculations within the LV2 plugin to be
similar to Jack and ALSA as much as possible. The special challenge with
LV2 is that the Host may invoke the plugin with variable sized processing
requests, even zero sized, and not necessarily a power of two.

The previous implementation used to perform some very specific processing
to accommodate this challenge. However, after careful investigation, I consider
this elaborate additional functionality pointless and not worth the complexity.

(1) the original implementation evaluated the timing information passed by
    the LV2 host for MIDI events. It attempted to "fast-forward" the audio
    calculations to the point behind the given timing of a MIDI event, presumably
    to generate more precise note timings. However, since the complete Yoshimi
    synth engine always starts a new note's envelope at the beginning of a full
    processing buffer, this very complex code can only ever make any difference
    for /some/ notes and under very specific and hard to reproduce circumstances.
    Addint to thise downside, the existing logic was rather complex and hard to
    verify, and turned out to be slightly imprecise or incorrect in corner cases.
    Especially dropping any MIDI event with a timing behind the current calculation
    request is insufficient; to properly implement such we'd need to maintain
    our own internal queue of not-yet-processed MIDI events.

(2) the original implementation used to let the SynthEngine calculate its results
    directly into the output buffer, which indeed saves the overhead of a data
    copy from the synth data buffer into the output buffer. However, these savings
    can be considered small in comparison to the actual synth calculations,
    since modern CPUs are quite good at moving data around. And, most importantly,
    on the downside this optimisation forces to run the SynthEngine on variable
    calculation chunk sizes, incurring a lot of complexity and even causing
    sonic glitches due to some hard to track-down mismatch in handling this
    variable length runs correctly. Which, after all, is in contradiction
    to the original design of the engine.

To put these changes into context, please consider the default case of processing,
which is using the ALSA or Jack engine. In both cases, samples are copied from the
synth buffer into output (in case of ALSA even with interleaving and endian swapping),
without this causing any noticeable performance problems. Please also note that both
the ALSA engine and the Jack engine collect all MIDI events encountered up to the
next processing call and thus always align these notes to the beginning of the next
calculation chunk. To further put this into context, consider the fact that a buffer
length of 128 audio samples at 48kHz corresponds to ~3ms, which is about the minimum
level of timing differences a experienced and trained musician is just able to discern.
And there still is the option to run the IO-system (ALSA, Jack) with only
64, 32 or maybe even 16 audio frames, given apropriate hardware.