Settings: retain calculation buffer size as main config parameter for now
This reverts a change to the UI done in
76c9ac

We just recently found out that the calculation buffer size has an impact
on the generated sound (which it should not have). In the light of this finding,
we need to retain the calculation buffer size as a user adjustable setting
for all I/O systems. In theory, it could be limited to ALSA, because Jack
and LV2 impose their own period size anyway. However, with that not yet
clarified impact on the generated sound, this setting should not be
chosen automatically in any case.