Fix a Segfault when loading LV2
The reason is a re-entrance from within the SynthEngine ctor,
which invokes the Config-ctor, which in turn invokes functionality
on the SynthEngine instance.

This problem was present since eda3f3c10
However, it did not cause any harm, since the invoked functions
happened to work without problem on the only partially initialised
SynthEngine instance.

With the change "Refactoring: extract a Singleton TextMsgBuffer..."
we introduce a TextMsgBuffer singleton, and SynthEngine now holds a
reference -- which at that point is not properly initialised.

The Fix is obviously not to invoke functionality on a partially initialised
SynthEngine instance; rather we can invoke it from oshimiLV2Plugin::instantiate()