Refactoring: turn the CmdInterpreter into a private component
There is a common guideline "favor composition over inheritance".

This means, base-classes and inheritance should only be used to express
implementation of an interface, or, generally speaking a "is-a" relationship.

Inheritance should not be used to get easy access tosome implementation.

If we just use a private implementation helper, as in the case here the
CmdInterpreter, it is better to place it explcitly as a private member
into the class. This makes the access and manipulation of encapsulated
state more explicit and clear.

Moreover we can now move the whole CmdInterpreter class into namespace cli.