Fix preset clipboard not copying/pasting
When copying to the preset clipboard, the XML would start with a
newline, which ultimately caused pasting not to do anything.

I'm fairly certain the original uses of strcmp were meant to test if
name started with "?xml" or "string" respectively, but strcmp compares
strings of arbitrary length lexicographically and a string always
compares lexicographically greater than a prefix of itself.

The magic numbers could be avoided by using sizeof(a_string_literal),
but this requires defining a macro or else it returns the size of a
pointer. I once discovered you can also use C++14 constexpr magic to
define a zero-overhead wrapper class for string literals that holds a
compile-time evaluated length field, but that's definitely overkill just
for this.