Fix ternary comparison with array
freqbasedmod is a field array, which coerces to a pointer that is never
nullptr, which coerces to true.

There's a similar line in computeVoiceOscillatorFrequencyModulation
which acts as the template for this change.

Frustratingly, this code only ever gets run when a note is played and an
ADSynth voice is enabled which is modulated, and whose oscillator source
is another voice which is modulated *by* yet another voice which is
modulated with either ring or morph modulation. For other modulation
types, the ternary condition is supposed to evaluate true so the fix
doesn't matter, and if the conditions outlined above are not met this
code doesn't run at all. But the bug is fixed regardless, and in this
edge case I can confirm the change affects the sound of such specially
crafted instruments.

If someone manages to actually create an instrument which relies on this
fix to sound good, I'd be amazed.