Remove dead code/duplicate logs in CMakeLists.txt
When REQUIRED is passed to find_package, pkg_check_modules, etc. CMake
aborts with an error if the package isn't found, so if execution reaches
the branch, *_FOUND is always true.

CMake also logs when a package is found, so logging this manually is
redundant and results in logs like:

-- Checking for module 'lv2>=1.0.0'
--   Found lv2, version 1.16.0
-- Found lv2 package 1.16.0