Fix OOB access in Data2Text.cpp and MiscGui.cpp
In both cases, part was declared and initialized without checking that
npart is < NUM_MIDI_PARTS, even though its value is only used in
branches where such a check has been done.

In Data2Text.cpp the shuffling around of the if tree to fit in the part
variable is a little awkward, but the massive if chain should probably
be refactored anyway.