Fix cast-to-unsigned overflow in Microtonal.cpp
octavesize is a size_t and division where either operand is unsigned
casts the result to an unsigned type, producing wildly wrong values when
(nt - ntkey) is negative.

A typedef that is signed and guaranteed to have the same width as size_t
doesn't seem to exist in a portable fashion, so int is used instead.

This could also be solved by changing octavesize to a signed type, but
there are a bunch of for loops with size_t counters that need to be
changed as well to avoid compiler warnings about mixed sign comparisons.

Why is octavesize size_t anyway? MAX_OCTAVE_SIZE is only 128, and
indexing arrays with signed values is perfectly valid.