Silence "using implicit namespace" warning
GCC doesn't complain, but Clang does.

Here, the std namespace is imported before any #includes that would
define it. This doesn't cause problems with GCC or Clang, but may not
work correctly with compilers that don't treat std specially.

Note that "using namespace std;" is considered an anti-pattern by some.