Update if() formatting in CMakeLists.txt
Repeating the if() expression in else() and endif() is no longer
necessary in CMake, and only supported for legacy reasons, so I have
removed instances of this. elseif() also exists, so I made an if()
pyramid much more readable.

This is true at least as far back as the minimum supported version,
CMake 3.0.