Fix UI bug which caused occasional noise in the oscillator.
It was a seemingly small issue, but with a quite substantial fix: The
UI and the audio thread were both using the OscilGen object to
generate samples and harmonics. This object is not thread safe, and
cannot easily be made thread safe either.

The solution is to split OscilGen into OscilGen, which is now private,
and OscilParameters, which is shared. The UI now creates its own
OscilGen object instead of using the one in ADnoteParameters.

Signed-off-by: Kristian Amlie <kristian@amlie.name>