Fix two bugs that reset harmonics to zero.
One was that the OscilGen constructor would set all values to
defaults, even the parameters. It should only set its own values to
zero, or it will overwrite the ones that were loaded.

The other was that we had an optimization to avoid calling
getbasefunction() from prepare() if certain parameters hadn't changed.
This isn't needed anymore because of the checkUpdated() call before we
call prepare(), and in fact is wrong. If the same patch was loaded
twice, none of those values would be different, but checkUpdated()
will still return true because we need to regenerate the base function
after loading a new patch. So just remove the check altogether.