Fix group phase randomness not working.
The intention behind the integer cast was probably to cast the whole
expression to int, but instead it cast only the first part, which has
to be float. It's not necessary to cast the whole expression, since it
will be done implicitly by the compiler when assigning to outpos, so
just remove it.