Fix shift-by-negative in bitFindHigh()
When value == 0, bit would eventually decrement to -1 before being used
as a shift argument.

I also removed reliance on int being 32-bit, and changed the method used
to test each bit. Some compilers may find it easier to translate (x & (1
<< y)) as a bit-test instruction, although this is not guaranteed.
Currently, on my machine, Clang generates a bit-test instruction when
using (x & (1 << y)), but not when using (value >> bit == 1).

GCC generates a shift and comparison in either case. GCC uses shrx and
cmpl for the right-shift method, and shlx and testl for the left-shift
method. GCC also generates a movl instruction (for the 1-literal) for
the left-shift method, but this is done outside the loop and ILP may
mean this has zero performance cost.

The above musings on performance are just splitting hairs, though; if
bitFindHigh is used anywhere performance is a concern, the correct
solution is to use a count-leading-zeroes intrinsic. On x86, the
behaviour of bitFindHigh is (except for returning 0xff on zero input)
exactly equivalent to the bsr (bit scan reverse) instruction. Using this
instruction instead of a loop could be as much as 100 times faster,
depending on branch prediction, etc.