Fix elusive bug in SUBnote filter creation, resulting in NaN.
It happened because the freq variable in the filter was set after the
yn1 and yn2 values were initialized in the initfilter function. In the
vast majority of cases it made no audible difference, because it is
only a starting value and will almost immediately drift towards the
correct value. However, very occasionally it would result in being set
to NaN, and this would quickly spread through the audio chain,
resulting in silence.

Thanks to floating point exceptions for tracking this one down, not
sure if I would have found the issue without them!