Fix extreme amplitude at the beginning of certain notes.
The problem was due to FMnewamplitude being initialized wrongly too
early. The initParameters() and computeNoteParameters() are meant to
initialize a fresh note, and update the parameters of an existing
note, respectively. FMnewamplitude needs a starting value, and hence
this is being set in initParameters(). However, it depends on the
value of FMVolume, which is set later in computeNoteParameters(). So
instead of calling those two functions in sequence, call
computeNoteParameters() from initParameters() at the correct moment so
that we can update FMnewamplitude afterwards.