Remove redundant !container.empty() checks
The conditional in if(!vec.empty()) vec.clear() is a no-op and serves
only as clutter.

list.remove() is a no-op on an empty list.

In if(!str.empty() && str != "notempty") while(!str.empty()) { ... },
the first emptiness check serves only to optimize for the case where str
is empty, at the expense of an extra conditional when it isn't. The
observable behaviour is the same in either case. If the compiler decided
to inline the != function, the generated assembly would probably be the
same, but both Clang and GCC insert a function call, which in the
empty-string case then presumably short-circuits on a loop condition
like if(i < str.size()).

In all these cases it is no more efficient to check for emptiness first.
Any good compiler with optimizations on will likely generate identical
assembly in either case.