Remove redundant clear()s
Both flatbankprgs and roots are std containers and members of
YoshimiLV2Plugin and Bank respectively.

They both get default-constructed when their containing class is
constructed, leaving them empty, thus a subsequent clear() is redundant.

The destructor for std containers usually (probably always) *is*
clear(), so calling it on members inside the containing class's
destructor is redundant.