Fix race condition in legato with very quick Note Off -> On sequence.
If a Note Off and On came in very quick succession, then the legato
would be applied to the not-yet-killed note which just had a Note
Off. However, this note would die soon after, leaving silence where
there should be a new note. Fix this by making sure we reset the
legato fading status upon a Note Off, which will trigger a new note
instead upon the next Note On.