Refactor ADnote legato
Overhauls the ADnote legato logic, which fixes legato notes losing phase
sync and makes legato crossfade work correctly. Crossfading is ignored
when portamento occurs, as there shouldn't be any need. That never made
sense anyway, even if the default 5ms fade time made it hard to notice.

SUB and PAD synths are not affected, but should be able to be overhauled
in a similar way. The code reuse between the synths should probably be
fixed (ZynAddSubFX already did this some time ago), but one thing at a
time.