Implement copy constructors for filters
Needed to implement copying notes, which is needed to implement legato
crossfading in a less buggy and (arguably) less kludgy way.

Virtual copy constructors are impossible, so instead the Filter_
superclass gets a clone() function, which takes a pointer and returns a
pointer to a newly allocated copy. For people who care about doing
things the Right Way, this is an established Pattern, but nonetheless
sheds light on one of C++'s shortcomings.

Conveniently, Filter (not Filter_) has no virtual methods but allocates
itself a Filter_, so Filter can be copy-constructed the normal way.