Improve dormant legato note behaviour
Previously, noteout() had to bail out if the note was a silent legato
note, and had to memset() its output buffers to avoid weirdness. Now,
ComputePartSmps() won't call noteout() on silent legato notes.

The ready() function used to be a regular data member that didn't really
make sense. It got set to 0 at the start of initialization and then set
to 1 at the end, but since this was all in the same function, any good
compiler would ignore the 0 assignment entirely. Regardless, it isn't
safe (or even possible, really) to rely on the ordering or timing of
non-atomic operations. The variable was only ever accessed from a single
thread anyway, so I'm really at a loss as to what the original intention
was.