Move legato fade step size to SynthEngine
This is only dependent on the sample rate, so it makes much more sense
to store it alongside the sample rate. I also made the math hopefully a
little clearer, and it doesn't need -ffast-math to compile to a single
division at runtime.

Although, seeing as these are normally only assigned to once for the
entire lifetime of the program, it's not exactly a *useful*
optimization, but three divisions looks tidier and clearer than mixing
division and multiplication in my opinion.