Partially fix PADnote portamento
PADSynth uses multiple wavetables, selecting the one with a base
frequency closest to the actual note frequency to (presumably) avoid
aliasing. This makes portamento difficult to do right, as to avoid
aliasing, you have to deal with the correct "current" table position
being different depending on frequency, and it looks like when
generating the tables the phases are randomized per-table, which would
need to change to allow smoothly transitioning from one table to another
anyway.

The alternative is what this change effectively does, which is keep the
current table no matter what the new note frequency is. When sliding up
in pitch significantly, aliasing starts to be audible, and I guess
sliding down a significant distance results in effectively a
low-pass-filtered sound as the high frequency content that was above the
Nyquist limit at the higher pitch is now in the audible range, but
absent in the table.

Fixing this should be possible, but would require modifying PADSynth
slightly. I haven't attempted to do this yet, but I think changing the
table generator to use the same set of randomized phases for all the
tables and changing legatoFadeIn() to convert the current table position
to its equivalent based on the difference in the two tables' base
frequencies should work.