Remove an unnecessary heap allocation of FFTwrapper
FFTwrapper is only 40 bytes on a 64-bit system, so there's no need to
allocate it on the heap. It already stores the bulk of its data
internally behind pointers.

There's a couple of other places where FFTwrapper is allocated with new,
but removing the indirection there would require either moving the
construction of the FFTwrapper into the containing class's
initialization list or implementing a null state for FFTwrapper (can
just use one of its internal pointers as a sentinel, so no need to make
the class fatter). It's probably not difficult, but there isn't much
gain to doing this.

Alternatively, the C++17 standard has std::optional<T> which might suit
this kind of situation better than supporting an uninitialized state in
the class itself, but it adds a byte + padding in overhead even when the
wrapped class could represent a null state internally.