Fix zombie legato notes
When I reimplemented legato, I inadvertently made the posb notes stay
alive *forever*, rather than only until the note is released and fading
out is finished. Oops.

I'm not hugely happy with the method used for this, but it works well
enough and is a relatively minor change. A potential improvement to
legato is doing away with the whole "dormant note" thing entirely. It
mainly only exists as a holdover from the old legato code, and because
it lets already-fading legato notes be replaced if necessary, rather
than potentially ending up with three or four notes fading at the same
time (an edge case with 5ms fade duration, sure) which would require
special logic to keep the sum of the fade values at 1.0.