Fix faulty array copies for legato notes with PWM
The ADnote copy constructor failed to take into account that Unison_size
in adpars isn't the value used for array sizing when pulse-width
modulation is used for a voice, so it copied only the first half of each
array that depends on unison count for its size, and worse, only
allocated half as much space as the rest of the code expects, leading to
out-of-bounds accesses.

Confusingly, unison_size[nvoice] *is* the value used for array sizes and
is set in ADnote::construct().