Clean up parent logic in ADnote copy constructor
Kristian changed the constructor to specify the true origVoice when
calling itself on subvoices, rather than specifying only the immediate
parent (i.e. "this"). This was intended as a bug fix, but didn't
actually fix anything as the existing code already checked for
parent->origVoice.

However, in hindsight it's simpler to have it take origVoice itself as
an argument, especially as subvoices are never (and probably should
never be) copy-constructed by anything other than the root voice's copy
constructor.

This commit removes the now redundant logic that checks if the specified
parent has an origVoice and renames the parent arg to origVoice_ for
clarity and consistency.