Clarify reference samplesize in PADnoteParameters
It's now slightly more obvious what the magic number is at a glance. I
could've moved this to a constant in the header, or perhaps just
elsewhere in the source file, but it's only used in one place (unless
you count it being equal to synth->samplesize_ref_f by coincidence) and
is unlikely to change in the future.