New algorithm for FM modulation in sub voices with parent modulation.
This one has been bugging me for a while. The old algorithm works well
as long as the frequency and amplitude of the modulator are stable.
However, it has a problem with artifacts when either of those are
changing (especially frequency), such as when using an envelope or an
LFO.

The problem stems from the fact that for each block, the algorithm
always starts at zero phase and works its way to the current phase
using the frequency which is in effect in that block. But this result
can be quite different for the next block if the frequency is
different. The larger the phase offset, the more of this frequency
difference will build up, causing a discontinuity.

The fix is to instead consider the *current* phase to be the center,
and always work backwards and forwards from there. The difference is
easy to hear in a sound with two layers of FM modulation, both with a
frequency or amplitude envelope or LFO.

This results in a rewritten algorithm which:

* Performs better (we start at the right phase instead of having to
  work our way there).

* Requires less memory (the table caching the samples between the zero
  phase and the current phase is no longer necessary).

* Is easier to understand.

* Doesn't break when using envelopes or LFOs on the modulators.

It doesn't get better than that!

Signed-off-by: Kristian Amlie <kristian@amlie.name>