Don't treat failure to write to GUI thread as an error.
This happens all the time when running as an LV2 plugin, because the
UI is usually not open unless working specifically on the Yoshimi
sound. Most of the time it is just doing its job in the background,
without UI. When this is the case, the events to the GUI are never
consumed and these messages start appearing by the hundreds. This can
be confusing, but it also tends to generate xruns, because printing is
not RT-safe in general. Ignoring these errors is harmless, however,
because the UI will consume them as soon as it is opened.