Optimize: Avoid running a full note update when only one part has changed.
LFOs and Envelopes are independent and don't need any changes to the
rest of the note data to provide live feedback.

To enable this for filters we need to move their parameter calculation
into computeWorkingParameters, which is called on every cycle, but we
are only talking a small number of calculations, which is nothing
compared to calling computeNoteParameters. This also allows us to get
rid of several working variables, since they are only needed once.

Signed-off-by: Kristian Amlie <kristian@amlie.name>