Fix detune accumulation bug in SUB synth when altering live voice.
Switch the name `initialfreq` to `basefreq`. This has no functional
impact, but is merely to match the naming in ADsynth; less surprises
for the reader. And then introduce `notefreq` to mean the frequency
which with all detuning applied. ADsynth has no equivalent of
notefreq, it is updated on every cycle. Also rename `setBaseFreq` to
`computeNoteFreq` and remove the argument, to make it more clear that
this function just updates the internal state.

After the renaming, make sure that notefreq is used in all places
where we need the detuned note frequency, and basefreq in all places
where we need the original, unaltered frequency.

Signed-off-by: Kristian Amlie <kristian@amlie.name>