Fix crash in legato mode when playing disabled, external voice.
If you import an external voice in ADnote in legato mode, then disable
the external voice, and try to play a legato progression, Yoshimi
crashes. This happens because VoiceOut should be created for all
notes, even disabled ones, but legato notes have been an exception
because they use the copy constructor, where it's missing. Fix by
reordering so that VoiceOut is created also in the copy constructor.

Signed-off-by: Kristian Amlie <kristian@amlie.name>