Make it possible to adjust Echo delay without losing tail.
As part of this, reimplement InterpolatedParameter to be more
generic. The new implementation is called InterpolatedValue. The main
motivation behind this change is to be able to cross fade not just
between two values, but between two array positions. This requires
that you know the start and end value exactly, which will be used to
index into the array. Then you grab the cross fade factor from
InterpolatedValue instead of the interpolated value, and use that to
cross fade between the two.

The new class also has some slightly different behaviors from the old
one:

* The new default cross fade length is 50ms instead of 10ms. This
  produces better results with Echo.

* New target values that arrive during the cross fade are postponed
  until the next cross fade. In the old implementation the cross fade
  would restart every time a new value arrived. Both have tradeoffs,
  but the main motivation is to have an upper bound on how long it
  will take for a value to be reached when new target values arrive
  very rapidly. For the new implementation it is at most two
  crossfades. For the old implementation it could, in extreme cases,
  be almost unlimited.

The first effect to use this is Echo, where the array access is needed
to access the wave form from X samples ago.

Signed-off-by: Kristian Amlie <kristian@amlie.name>