Fix Equal Temperament variation when voice is being imported.
When the voice is imported, the 440Hz is not used, the frequency is
taken from the importing voice. However, it should still be possible
to use the Eq.T. wheel if 440Hz is enabled in both voices.

With this fix it is even possible to use Eq.T. without 440Hz enabled,
if the importing voice has it disabled. The effect is then added to
the existing 1x equal temperament from the importing voice.

Signed-off-by: Kristian Amlie <kristian@amlie.name>