TestInvoker: initial draft
The TestInvoker within Yoshimi shall be used for acceptance testing
of the sound calculation without any MIDI or audio I/O.

When such a test is invoked, first send a CLI commmand to stop any ongoing sound production,
so we can basically ignore the actual Audio / MIDI backend (or use a null backend).
Then a suitably sized buffer will be allocated, the SynthEngine will be reseeded,
and then the sound calculation will be invoked from the test function.

NOTE: deliberately we invoke the sound calculation direct (in CLI thread) and blocking

With the next commits, the calculated sound data will be  written into a raw file --
which can then be opened and processed from the Testsuite / Testrunner.