Implement BPM synchronization for all Effects.
All of them except Echo use the Beat values to track where in the LFO
cycle they are, which is relatively exact, and more than good enough
for these effects. Echo is not based on an LFO, but on a delay time,
and thus uses the BPM instead. For Jack and LV2, this value is exact,
but for Alsa, it is calculated from the MIDI clock, which is quite
susceptible to jitter, which is very audible in the Echo effect. Quite
a lot of smoothing effects are in place to hide this, including an
average of the last x calculated median BPM values, as well as
refusing to update the delay time for changes smaller than 2%.

Signed-off-by: Kristian Amlie <kristian@amlie.name>