Fix sporadic crash in Echo effect.
There are two issues here: First of all, we should clear the whole
delay line in cleanup(), not just the part we can hear. This is
important if the delay is later changed, which may reveal old audio
which lingers there. But more importantly, it avoids using dl and dr,
which may be uninitialized at that time.

But that is also a problem in itself: dl and dr should be initialized,
so make sure they always are.

Signed-off-by: Kristian Amlie <kristian@amlie.name>