BugFix: Corrupted instrument Copyright and Comments entries.
This was initially caused by a change to FileMgrFuncs loadText(), intended to allow for MS line endings and other oddities.
It had succeeded but in a way that messed with the formatting applied to the above entries.
The new loadText() code solves the problems for both.

At the same time, the text formatting had a known bug where it would mess up repeated re-formatting, so this has been reworked too.

Finally, the formatting *request* has been taken out of the GUI and put in the main code.