Completed latest banks update
Previously on loading a bank each instrument file was scanned for
its type, engines and whether yoshi or zyn.

Instrument name and this data is now stored in the banks file itself
so only the file's existence is checked. This is *much* faster.
When run for the first time it will use the old slow method.

Any runtime changes to individual instruments are checked as before.
The same is true for instruments in imported banks.

The Updated banks file is saved on exit.
The new structure is backward compatible.

Additional corrections:

On importing banks, instruments files with a second hyphen were
incorrectly identified.

Importing banks with instruments having a prefix of zero now
installed as max (159). Previously ignored or marked as invalid.

Instruments added to banks by external means previously disrupted
all the recognised ones - now ignored.

Deleting a bank didn't remove the directory itself.